refactor: Name, Position 클래스 제거 후 기본형으로 변경
