--- src/main/java/racingcar/domain/Name.java
@@ -1,23 +1,0 @@
-package racingcar.domain;
-
-public class Name {
-
-    private String name;
-
-    public Name(String name) {
-
-        validateName(name);
-        this.name = name;
-    }
-
-    private void validateName(String name) {
-
-        if(name.length() > 5) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public String getName() {
-        return name;
-    }
-}
