--- src/main/java/racingcar/domain/Position.java
@@ -1,15 +1,0 @@
-package racingcar.domain;
-
-public class Position {
-
-    private int position;
-
-    public void move() {
-
-        position++;
-    }
-
-    public int getPosition() {
-        return position;
-    }
-}
